fix(storage): resolve scan_id binding bug, improve error handling
- Add scan_id field to StoredFinding, fixing insert_finding which
  incorrectly bound affected_asset.identifier as scan_id
- Replace dummy finding on deserialization failure with proper error
  propagation
- Add tracing::warn for unknown scan/finding status defaults
- Replace Debug format with Display for ScanStatus DB storage
- Split list_findings into per-scan and list_all_findings methods
- Add offset/limit pagination to list_findings
- Remove unused StoredUser/UserRole types
- Add pg_trgm threshold cast in find_duplicate_findings query

+++ crates/slapper/src/storage/queries.rs
@@ -21,7 +21,7 @@ impl QueryBuilder {
         "SELECT f1.id, f2.id as duplicate_id
          FROM findings f1
          JOIN findings f2 ON f1.id < f2.id
-         WHERE similarity(f1.finding->>'title', f2.finding->>'title') > $1"
+         WHERE similarity(f1.finding->>'title', f2.finding->>'title') > $1::float"
     }
 }
